Added support for the Windows 10 platform.
This supports MRT icons for Windows, which is Windows' version of its resolution scaling.
The Windows resource array defines all the scales needed, and those files are created with the required width and length. Ther MRT standard is implemented at Cordova with only one config entry for all of the various scales built, and this is achieved by adding a 'target' attribute, and ensuring that, for the Windows platform, that only images with a target attribute are written to the config file.

Code changes:

- A new Format.NONE type indicates that the file should not be written
- A new Platform.WINDOWS type defines the Windows platform
- A new WindowsTargets enum holding all the Windows targets
- Added 'target' to the types that define and pass around the source images.
- Support through the code for the windows platform

Tests are included.

Results in the config look like this:

        <icon src="res/windows/icon/SmallTile.png" target="Square71x71Logo" />
        <icon src="res/windows/icon/Square150x150Logo.png" target="Square150x150Logo" />
        <icon src="res/windows/icon/Wide310x150Logo.png" target="Wide310x150Logo" />
        <icon src="res/windows/icon/Square310x310Logo.png" target="Square310x310Logo" />
        <icon src="res/windows/icon/Square44x44Logo.png" target="Square44x44Logo" />
        <icon src="res/windows/icon/StoreLogo.png" target="StoreLogo" />
        <splash src="res/windows/splash/Splash.png" target="SplashScreen" />